momentum-ai-universal/ai-service/requirements.txt
# Core FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
python-multipart==0.0.12
python-dotenv==1.0.1

# Database and vector stores
supabase==2.10.0
httpx==0.27.2
faiss-cpu==1.7.4
pymilvus==2.3.6
pgvector==0.2.5

# LLM and embeddings
nomic==2.0.3
llama-cpp-python==0.2.6
transformers>=4.41.0,<5.0.0
huggingface-hub>=0.20.0
sentence-transformers==3.3.0
langchain==0.1.9
langchain-community==0.0.24

# ML and data processing
numpy==1.26.4
scikit-learn==1.5.2
pandas==2.2.3

# Visualization and monitoring
streamlit==1.31.1

# Additional utilities
aiofiles==24.1.0
requests==2.31.0 